% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{assessment_plot}
\alias{assessment_plot}
\title{assessment_plot}
\usage{
assessment_plot(x)
}
\arguments{
\item{x}{output from assessment_table_builder}
}
\value{
prints a ggplot object
}
\description{
Called from wthin visualise_assessment.
Plots the results of an assessment in a stacked bar chart.
}
