% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{assessment_table_builder}
\alias{assessment_table_builder}
\title{assessment_table_builder}
\usage{
assessment_table_builder(x, cols = colour_lookup)
}
\arguments{
\item{x}{a data frame created by do_assessment}

\item{cols}{colour lookup table}
}
\value{
a summary data frame
}
\description{
Called from wthin visualise_assessment.
Builds a summary table from as assessment data frame
}
