% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{do_assessment}
\alias{do_assessment}
\title{do_assessment}
\usage{
do_assessment(
  variables,
  targets,
  thresholds,
  smoothed_trend,
  short_term,
  long_term,
  target_term
)
}
\arguments{
\item{variables}{String containing the names of the variables}

\item{targets}{output from load_process_metadata}

\item{thresholds}{output from load_process_metadata}

\item{smoothed_trend}{output from get_smoothed_trend}

\item{short_term}{set by user, number of years of data to use for short term assessment}

\item{long_term}{set by user, number of years of data to use for long term assessment}

\item{target_term}{set by user, number of years of data to use for target assessment}
}
\value{
a list of three dataframes, one for each assessment
}
\description{
Performs a short term, long term and target assessment on a list of variables.
}
