% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{get_smoothed_trend}
\alias{get_smoothed_trend}
\title{get_smoothed_trend}
\usage{
get_smoothed_trend(x)
}
\arguments{
\item{x}{A dataframe (dat_list) output by load_process_data}
}
\value{
A list of smoothed trends
}
\description{
Apply a loess smoother trend to your data, predicting value from year.
Uses default setting span = 0.75 which is not appropriate for small data sets, warnings are produced
and returned to the console.
}
