% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{load_process_data}
\alias{load_process_data}
\title{Load and process data}
\usage{
load_process_data(filename)
}
\arguments{
\item{filename}{the name of the file which the data are to be read from.
Each row of the table appears as one line of the file. If it does not
contain an absolute path, the file name is relative to the
current working directory, getwd().}
}
\value{
Two dataframes.
}
\description{
load_process_data takes an input string, reads a file in table format, creates a data frame
from it and parses it to return two dataframes: dat_list and variables. These outputs are
assigned directly to the global environment.
}
