% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{load_process_metadata}
\alias{load_process_metadata}
\title{load and process metadata
load_process_metadata takes an input string, reads a file in table format and creates a data frame from it,
and parses it to return three dataframes: targets, thresholds and goal_indicator_lookup. These outputs are
assigned directly to the global environment.}
\usage{
load_process_metadata(filename)
}
\arguments{
\item{filename}{the name of the file which the data are to be read from.
Each row of the table appears as one line of the file. If it does not
contain an absolute path, the file name is relative to the
current working directory, getwd().}
}
\value{
Three dataframes.
}
\description{
load and process metadata
load_process_metadata takes an input string, reads a file in table format and creates a data frame from it,
and parses it to return three dataframes: targets, thresholds and goal_indicator_lookup. These outputs are
assigned directly to the global environment.
}
