% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{save_smoothed_trend}
\alias{save_smoothed_trend}
\title{save_smoothed_trend}
\usage{
save_smoothed_trend(x)
}
\arguments{
\item{x}{a list of data frames containing teh time series of each variable}
}
\value{
saves a .png to file. The file name is relative to the
current working directory, getwd().
}
\description{
Visualises the smoothed trend for a variable
}
