% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{target_assess_this}
\alias{target_assess_this}
\title{target_assess_this}
\usage{
target_assess_this(x, targets, smoothed_trend, term = 5)
}
\arguments{
\item{x}{The name of the variable as a string.}

\item{targets}{output from load_process_metadata}

\item{smoothed_trend}{output from get_smoothed_trend}

\item{term}{set by user, number of years of data to use}
}
\value{
a list containing the first and last values in the series, the rare of change, the
number of years until target is reached and the result of the assessment.
}
\description{
Called from within do_assessment.
Performs an assessment of change against a target on an indicator or variable.
}
