% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{visualise_assessment}
\alias{visualise_assessment}
\title{visualise_assessment}
\usage{
visualise_assessment(
  classification = "Primary.goal",
  group = "Clean air",
  x,
  myLab = "assessment"
)
}
\arguments{
\item{classification}{the column name from which to subset the assessment by}

\item{group}{the factor level from within classification by which to subset the assessment}

\item{x}{output from do_assessment}

\item{myLab}{string specifying the x axis label}
}
\value{
prints a ggplot object
}
\description{
ties assessment_table_builder and assessment_plot together to visualise the
results of as assessment.
}
