% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{visualise_assessment_dot}
\alias{visualise_assessment_dot}
\title{visualise_assessment_dot}
\usage{
visualise_assessment_dot(
  x,
  classification = "Group",
  group = "Atmosphere",
  myLab = "Target assessment"
)
}
\arguments{
\item{x}{output from do_assessment}

\item{classification}{the column name from which to subset the assessment by}

\item{group}{the factor level from within classification by which to subset the assessment}

\item{myLab}{string specifying the x axis label}
}
\value{
prints a ggplot object
}
\description{
visualises the result of a target assessment as a dot plot.
}
