% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{years_until_target_reached}
\alias{years_until_target_reached}
\title{years_until_target_reached}
\usage{
years_until_target_reached(
  rate_of_change,
  final_value,
  temp_target,
  temp_target_trend
)
}
\arguments{
\item{rate_of_change}{output from trend_assess_this or target_assess_this}

\item{final_value}{output from trend_assess_this or target_assess_this}

\item{temp_target}{output from target_assess_this}

\item{temp_target_trend}{output from target_assess_this}
}
\value{
number of years (integer)
}
\description{
Called from within target_assess_this.
Calculates the number of years until the target is reached for a single indicator,
using the formula:
$$ years = abs(log(final_value/temp_target)/log(1 + rate_of_change))
years = abs(log(final_value/temp_target)/log(1 + rate_of_change))
see https://intl.siyavula.com/read/maths/grade-12/finance/03-finance-01 for more information
}
