% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pilot_assessment_functions.R
\name{trend_assess_this}
\alias{trend_assess_this}
\title{trend_assess_this}
\usage{
trend_assess_this(x, term = 5, thresholds, smoothed_trend)
}
\arguments{
\item{x}{The name of the variable as a string.}

\item{term}{Number of years to perform the assessment on.}

\item{thresholds}{output from load_process_metadata}

\item{smoothed_trend}{output from get_smoothed_trend}
}
\value{
a list containing the first and last values in the series, the rare of change and the
result of the assessment.
}
\description{
Called from within do_assessment.
Performs an assessment of change on an indicator or variable.
}
